cleanup code

+++ src/lib/io.h
@@ -1,7 +1,7 @@
 #ifndef IO_H
 #define IO_H
 
-int Io_printSpaces(int n);
+int Io_printSpaces(long n);
 
 
 #endif
