+++ src/lib/string.h
@@ -5,17 +5,17 @@
 
 typedef char *T;
 
-long String_equals (T x, T y);
-T String_new (char *s);
-T String_concat (char *s, ...);
+long String_equals(T x, T y);
+T String_new(char *s);
+T String_concat(char *s, ...);
 /*void String_append (T, T);*/
-T String_fromArray (int size, char *array[], char *sep);
-int String_size (T s);
-T String_toString (T x);
-int String_hashCode (T x);
-void String_print (T x);
+T String_fromArray(long size, char *array[], char *sep);
+long String_size(T s);
+T String_toString(T x);
+long String_hashCode(T x);
+void String_print(T x);
 //File_t String_printAsIs (File_t file, T);
-T String_toStrAsIs (T x);
+T String_toStrAsIs(T x);
 
 #undef T
